working on libphonenumber (not completed)
