+++ unitTest/unitTest.pro
@@ -8,12 +8,16 @@
 ################################################################################
 
 HEADERS += \
-    UnitTest.h
+    UnitTest.hpp \
 
 # +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-#
 SOURCES += \
-    UnitTest.cpp
+    main.cpp \
+
+# +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-#
+LIBS += \
+    -lprotobuf \
+    -lphonenumber \
 
 ################################################################################
 include($$QBUILD_PATH/templates/unitTestConfigs.pri)
-
